backend/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
PyJWT==2.5.0
pytz==2022.2.1
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
